+++ src/app.ts
@@ -3,7 +3,7 @@ import path from "path";
 import cors from "cors";
 import routes from "./routes/auth.route";
 import dotenv from "dotenv";
-import { dbConnection } from "./db/config";
+import { dbConnection } from "./db/config.db";
 dotenv.config();
 
 const app = express();
